+++ contrib/lualibs/mympd/50-util.lua
@@ -141,3 +141,29 @@ function mympd.remove_file(path)
   end
   return rc, errorstr
 end
+
+--- Checks arguments from the mympd_arguments global variable.
+-- @param tocheck Table of arguments with options to check
+--                Available options: notempty, required, number
+-- @return true on success, else nil
+-- @return Error message on failure
+function mympd.check_arguments(tocheck)
+  for arg, option in pairs(tocheck) do
+    if mympd_arguments[arg] == nil then
+      mympd.log(3, "Argument " .. arg  .. " not found.")
+      return false, "Argument " .. arg  .. " not found."
+    end
+    if option == "notempty" then
+      if mympd_arguments[arg] == "" then
+        mympd.log(3, "Argument " .. arg  .. " is empty.")
+        return false, "Argument " .. arg  .. " is empty."
+      end
+    elseif option == 'number' then
+      if mympd_arguments[arg] == nil or tonumber(mympd_arguments[arg]) == nil then
+        mympd.log(3, "Argument " .. arg  .. " is not a number.")
+        return false, "Argument " .. arg  .. " is not a number."
+      end
+    end
+  end
+  return true
+end